requirements.txt
Django==2.0.6
django-redis==4.9.0
PyJWT==1.6.4
